* src/etc/munge.init.in: Added stanza to service_start() to create   /var/run/munge dir if needed.  Fixes bug on Debian/Ubuntu when RAMRUN   is enabled.
Ideally, the code to check/create the /var/run/munge dir should be
in the daemon.  But the code for running munged as a non-root user is
currently in the init script (via the $USER var).  As such, the code
for creating /var/run/munge must be there as well; otherwise, the
created directory could have the wrong ownership.  Perhaps this code
will be moved into the daemon once it supports a configuration file.