Added fd_timed_* optimization to skip first poll.
Added optimization to skip the poll() on the first iteration of the
loop for nonblocking sockets.  This helps the case where the read or
write can complete immediately and in a single invocation, thereby
eliding the overhead of gettimeofday() + poll().  This is safe for
nonblocking sockets since the loop already handles EAGAIN to continue
with the next iteration of the loop.