Added fd_timed_* "do_skip_first_poll" parm.
Replaced call to fd_is_nonblocking() with "do_skip_first_poll" parm.
This eliminates the overhead of a function call and fcntl() at
the expense of an additional function arg to determine whether the
fd is nonblocking.  The old behavior can be preserved by passing
"fd_is_nonblocking(fd)" for the new arg.