Specified a min buf size for xgetgrent/xgetpwnam.
Specified a sensible minimum buffer size for xgetgrent() and
xgetpwnam().  On some systems, the values returned by sysconf()
seemed artificially low for groups with a large number of users.

On one of our systems, xgetgrent() returned ERANGE for a group with
109 users where _SC_GETGR_R_SIZE_MAX was 1024.  Increasing the buffer
to 4096 allowed the call to succeed.  And there was much rejoicing.