Added _random_seed_stir_callback() timer.
Created a new timer callback to add entropy to the PRNG.  The source of
this entropy is the microsecond portion of the gettimeofday() result.
The maximum microsecond value is just under 20 bits, so the actual
amount of entropy gathered from each stirring will be somewhat less
(somewhere between 0 and 19.93157 bits).

The delay between callbacks starts at 1 second, and exponentially
increases up to a maximum of 4 hours.  This allows for a vigorous
stirring of the entropy pool when the daemon is started.