detect whether an existing domain socket is in use
Create an advisory write-lock for detecting whether the unix domain
socket is currently in use by another munged process.  If this
exclusive lock is held by another running process, the pid for that
process can be ascertained.  This lock will be automatically released
when the process terminates.

Since this lock can determine whether the associated socket is in use,
an existing socket resulting from an abrupt termination can be cleaned
up without having to "force" the daemon.

The lockfile is named by appending ".lock" to the socket name.  This
ensures it is associated with a particular socket and munged process.

Fixes issue 13.