preserve xgetgrent() buffer size across scans
Instead of having the xgetgrent() buffer start with the system
recommended size on each new scan, preserve the final size at the
end of a scan.  This size will be used to create the buffer on the
next scan.  Thus, subsequent scans will generally not need to have
their buffer realloc()d.  And on systems where getgrent_r() cannot
be safely resumed after ERANGE, this will reduce the number of scans
that must be restarted.

Update issue 16
preserve xgetgrent() buffer size across scans