check _timer_is_timespec_ge() tv_nsec bound
Check if the _timer_is_timespec_ge() timespec tv_nsec values have
overflowed into tv_sec.  This replaces assert()s with a runtime check
to correct the timespecs if needed.  I'm not aware of a condition
or platform that can return a tv_nsec >= 1e9, but this is an added
check for correctness.