fix xgetgrent() got_eof test for FreeBSD
With the GNU implementation of getgrent_r() (identified by the autoconf
define HAVE_GETGRENT_R_GNU), the 4th arg (which I'll call "gbufp")
points to the caller-provided "struct group" buffer passed as the
1st arg.

Once getgrent_r() has processed all of the entries in the group file,
the Linux implementation returns ENOENT and sets "gbufp" to NULL.
In contrast, the FreeBSD implementation returns 0 and sets "gbufp"
to NULL.

This updates the "got_eof" conditional test to check for either.

Fixes issue 18.