fix xgetpwnam() to detect errors on FreeBSD/SunOS
While forcing an ERANGE condition during testing of the POSIX
xgetpwnam(), the error was misconstrued as "user not found" on FreeBSD
(9.0, 8.2) and SunOS (5.11, 5.10); in particular, xgetpwnam() returned
-1 with errno=ENOENT.  This is because getpwnam_r() returned ERANGE
with errno=0.  xgetpwnam() expected the error value to be set in errno
(which is the behavior on both Linux and AIX); however, FreeBSD
and SunOS return the error in the return value of getpwnam_r().
Linux actually does both (ie, return the error in errno and the
getpwnam_r() return value).

Consequently, the behavior of the HAVE_GETPWNAM_R_POSIX case of
xgetpwnam() was changed as follows:
  1. Coalesce the error number in errno onto rv if needed.
  2. Presume "user not found" for rv of 0, ENOENT, ESRCH.
  3. Presume error for rv of EINTR, ERANGE, EIO, EMFILE, ENFILE.
  4. Assume anything else is "user not found".

This seems to work reasonably well on the systems I have available
for testing.  The notable exception here is the case of OpenBSD.
OpenBSD (5.0, 4.9, 4.8, 4.7) sets errno=EACCES whenever getpwnam_r()
is called, regardless of whether it succeeds or fails for any reason.
OpenBSD 5.1 corrects this by never setting errno in getpwnam_r().
On OpenBSD, there is no way to differentiate between "user not found"
and any other error.

Notes for HAVE_GETPWNAM_R_POSIX getpwnam_r() behavior:

Success:
  Linux (3.x, 2.6.x):
    rv=0        errno=0
  AIX (5.2)
    rv=0        errno=0
  FreeBSD (9.0, 8.2):
    rv=0        errno=0
  OpenBSD (5.1):
    rv=0        errno=0
  OpenBSD (5.0, 4.9, 4.8, 4.7):
    rv=0        errno=EACCES
  SunOS (5.11, 5.10):
    rv=0        errno=0

User Not Found:
  Linux (3.x, 2.6.x):
    rv=0        errno=0         resultp=NULL
  AIX (5.2)
    rv=-1       errno=ESRCH     resultp=NULL
  FreeBSD (9.0, 8.2):
    rv=0        errno=0         resultp=NULL
  OpenBSD (5.1):
    rv=1        errno=0         resultp=NULL
  OpenBSD (5.0, 4.9, 4.8, 4.7):
    rv=1        errno=EACCES    resultp=NULL
  SunOS (5.11, 5.10):
    rv=0        errno=0         resultp=NULL

Error (ERANGE):
  Linux (3.x, 2.6.x):
    rv=ERANGE   errno=ERANGE    resultp=NULL
  AIX (5.2)
    rv=-1       errno=ERANGE    resultp=NULL
  FreeBSD (9.0, 8.2):
    rv=ERANGE   errno=0         resultp=NULL
  OpenBSD (5.1):
    rv=1        errno=0         resultp=NULL
  OpenBSD (5.0, 4.9, 4.8, 4.7):
    rv=1        errno=EACCES    resultp=NULL
  SunOS (5.11, 5.10):
    rv=ERANGE   errno=0         resultp=NULL