preserve xgetpwnam() buffer size across scans
Instead of having the xgetpwnam() buffer start with the system
recommended size on each new scan, preserve the final size at the
end of a scan.  This size will be used to create the buffer on the
next scan.  Thus, subsequent scans will generally not need to have
their buffer realloc()d.