change munge(1) --string arg to not append newline
Change the behavior of the munge(1) "--string" command-line option
such that a newline is no longer appended to the argument if a trailing
newline is not present.

Originally, a newline was added to ensure the unmunge(1) output would
be properly formatted.  Now, unmunge(1) has been changed to append a
trailing newline if needed when outputting the payload to a tty device.