change failed hostname resolution err to warning
The IPv4 address of the encoding host is embedded within the
credential metadata, but is informational and not required for
successful authentication.  gethostname() is called to obtain the
host name, and gethostbyname() is called to resolve its IP address.

Since the IP address here is not strictly required for authentication,
change the munged behavior so that a failed gethostbyname() is not
fatal.  Instead, log a warning message and use the IPv4 address
0.0.0.0.

In addition, change unmunge(1) to avoid calling gethostbyaddr() for
0.0.0.0.  The avoids the delay for its name resolution to slowly fail.

Fixes issue 20.