update SysV init script chkconfig/LSB header
Update the description to refer to the service instead of the script.

Add $local_fs dependency to Required-Start to ensure the key, seed,
socket, and logfile are available when starting the service.

Add $local_fs dependency to Required-Stop to ensure the seed, socket,
and logfile are accessible when stopping the service.

Add $network dependency to Required-Start to ensure the network
interfaces have been brought up so the origin IP address can be
obtained when starting the service.  This IP address will be encoded
into the credential metadata.