move SysV init script VARRUNDIR def to top
Move the VARRUNDIR variable for specifying the location of the
/var/run/munge directory to the configuration block at the top of
the SysV init script.  This variable is used in service_start() to
create & chown that directory, if needed, when the service is started.

Normally, the /var/run/munge directory is created when MUNGE is
installed.  However, on systems setting RAMRUN in rcS, the /var/run
directory is placed on a tmpfs filesystem that is cleared at boot.
Setting VARRUNDIR in the init script ensures that directory will be
re-created when munged is started.