update rpm spec %files for sysconfig & init.d
Replace the "%{_sysconfdir}/*/*" glob with separate entries for the
sysconfig file and the init.d file.

The entry for the init.d file now uses the _initddir rpm macro (or
the deprecated _initrddir rpm macro if _initddir is not defined).
This allows it to properly expand to either "/etc/rc.d/init.d"
(Fedora/RHEL) or "/etc/rc.d" (openSUSE).