cleanup "--mlockall" cmdline opt to munged
This adjusts the issue-22 patch from Michael Fenn (1a94f1db6b5c)
for adding the "--mlockall" command-line option to munged.

The call to mlockall() is moved out of main() and into the newly
added lock_memory().  In addition, EAGAIN is tested for, in which case
mlockall() is retried up to max_tries (currently 10 times) since this
error is presumably transient.  But note that mlockall(MCL_CURRENT)
consistently fails with EAGAIN on OpenBSD 5.2, although it appears
to work on earlier and later versions.

The error message for when mlockall() fails (or is not implemented,
as is the case on AIX 5.2) is changed from being non-fatal to fatal.
If the --mlockall option is specified, it either succeeds or the
daemon terminates.

The #ifdef condition _POSIX_MEMLOCK is replaced with HAVE_MLOCKALL
since autoconf tests are already used throughout.

Tested on:
- AIX 5.2 (ENOSYS)
- Debian 6
- FreeBSD 9.1
- GNU/kFreeBSD 8.1
- NetBSD 6.0
- OpenBSD 5.3, 5.2 (EAGAIN), 5.1
- SunOS 5.10