allow failure of domain socket lock to be forced
As of 6988416f2902, munged uses an advisory write-lock on a lockfile
to ensure exclusive access to its domain socket.  A fatal error is
generated if it cannot be obtained.  This lock prevents a subsequent
munged process from inadvertently stealing (ie, unlinking and
re-creating) the domain socket from a running munged process which
would be unaware that this theft had occurred.  It is automatically
dropped when the process terminates.

On some systems, these types of locks are not supported on NFS mounts,
and the fcntl() in set_file_lock() will fail with errno=EOPNOTSUPP.

This commit changes the behavior such that a failure to obtain
this lock can be overridden via the "--force" command-line option.
When "forced", a warning message will be logged when this lock attempt
fails, and process execution will continue.

Update issue 13
Allow the "--force" command-line option to override the error generated
when failing to obtain the advisory lock for the domain socket.
On some systems, these locks are not supported on NFS mounts.