fix HMAC validation timing attack vulnerability
The use of memcmp for HMAC validation can leak timing information,
thereby allowing an adversary to iteratively try HMAC values, time
the responses, and progressively deduce the correct HMAC value.

This attack would allow for the forging of a single MUNGE credential
which must be completed within the credential's TTL.  This would
allow an adversary to rewrite the embedded UID/GID to enable privilege
escalation via some other service using MUNGE for authentication.

This timing attack dates back to at least 2009 when Nate Lawson
reported on finding a security flaw in the Google "Keyczar"
cryptographic library that would allow an attacker to forge signatures
for data signed with a SHA-1 HMAC:

http://rdist.root.org/2009/05/28/timing-attack-in-google-keyczar-library/

Nate revisits timing-independent array comparisons in a later post:

http://rdist.root.org/2010/01/07/timing-independent-array-comparison/

This commit replaces the memcmp used for HMAC validation with an
implementation that compares the two memory regions in an amount of
time that is independent of their contents.  It is based on pseudo-code
from the above URLs.

Fixes issue 24.