fix /var/run/munge dir ownership during upgrade
Commit f1f1480e9534 changes the default ownership of the running
munged process to the "munge" user.  It also changes the ownership
of existing files in /etc/munge/ and /var/*/munge/ to that of the
"munge" user (unless you've defined USER in the sysconfig file).

Commit 0f89429d60d7 adds the %ghost directive to the /var/run/munge
directory in the rpm spec %files list.

Since /var/run/munge is now tagged with %ghost, the user and group
ownership are not updated during an upgrade.

This commit updates the ownership of /var/run/munge during an upgrade.