add timeout when stopping daemon via init script
The init script uses system-specific facilities where available to
control the service.  On systems where these standard facilities are
not available, it falls back on a common implementation.  For stopping
the service, this entails checking the process listing, sending a
SIGTERM, rechecking the process listing, and sending a SIGKILL if
needed.  After sending the SIGTERM, the process listing is immediately
rechecked -- oftentimes before the daemon has a chance to shutdown.
This can result in needlessly sleeping for 3 seconds (by default)
before sending a superfluous SIGKILL.

This commit adds a 1-second sleep between the SIGTERM and the recheck
of the process listing.  This delay appears to give the daemon ample
time to shutdown, thereby shortening the average delay from 3 seconds
to 1 second.