fix rpm failed dependency for shadow-utils
Commit f1f1480e9534 switched the default behavior to run munged
as the "munge" user.  The RPM spec file created this "munge"
user/group account, if needed, in its %pre script.  An RPM Requires
for shadow-utils was added; this package is found on current Fedora
and RHEL systems.

However, the shadow-utils package is not available on openSUSE.
Historically, this requirement has been fulfilled with pwdutils.
But openSUSE 12.3 replaces pwdutils with shadow:

https://www.suse.com/releasenotes/x86_64/openSUSE/12.3/#idm47363415597696

This commit updates the RPM Requires to use shadow-utils, shadow,
or pwdutils depending on the system.