remove munged segv_handler()
Remove the munged SIGSEGV handler.  This handler logs a message
indicating the daemon received a segfault before terminating with
exit(1).  But it is interfering with the systemd "Restart=on-abort"
behavior which only restarts a service if its process exits due to
an uncaught signal.  Since this log message is of minimal benefit,
the easy fix is to simply remove segv_handler().