move munge.pc AC template to src/etc/
Move the pkg-config source file for munge.pc to src/etc/.  This file
does not belong in the top-level directory.  It's a source file
to be installed so it belongs somewhere in the src/ hierarchy.
While it does not get installed in /etc, the src/etc/ dir seems a
more appropriate place for assorted odds and ends.