add support for systemd
At "make install", the munge.service systemd config will be installed
in /usr/lib/systemd/system/ or /lib/systemd/system/ (if either is
present), and the munge.conf tmpfiles.d config will be installed in
/usr/lib/tmpfiles.d/ (if present).

Note that Debian 7, Ubuntu 13.04, and some older versions of openSUSE
use /lib/systemd/system/.  But Fedora 17+, RHEL 7, and openSUSE 12.3
are using /usr/lib/systemd/system/ which is the value of the RPM
%{_unitdir} macro.

Socket unit configuration for the munged socket is not (yet) supported.

The top-level munge.spec file keeps the traditional SysV init script.
RPM specfiles for systemd init on Fedora and openSUSE are in extra/.

With systemd, configuration of the munge service should be done by
copying /usr/lib/systemd/system/munge.service to /etc/systemd/system/
and modifying the new file as needed.

References:
- http://0pointer.de/public/systemd-man/systemd.unit.html
- http://0pointer.de/public/systemd-man/systemd.service.html
- http://0pointer.de/public/systemd-man/systemd.exec.html
- https://fedoraproject.org/wiki/Packaging:Systemd
- https://fedoraproject.org/wiki/Packaging:Tmpfiles.d
- https://fedoraproject.org/wiki/Packaging:SysVInitScript?rd=Packaging/SysVInitScript
- https://fedoraproject.org/wiki/Packaging:ScriptletSnippets#Systemd
- http://en.opensuse.org/openSUSE:Systemd_packaging_guidelines

Tested on:
- Fedora 19, 18
- RHEL 7.0 Alpha3
- openSUSE 12.3

Fixes issue 11.