Always install pkgconfig munge.pc file
When installing the pkgconfig munge.pc file, the Makefile would
first check for the existence of the "$(libdir)/pkgconfig" directory.
If this directory did not exist, the Makefile would omit installing
munge.pc under the presumption that pkgconfig was not being used.

However, when $(prefix) is redefined, $(libdir) is effectively
redefined as well since it indirectly references $(prefix).
This likely results in the munge.pc installation being omitted.
And this omission results in rpmbuild failing with "File not found"
for the munge.pc file.

Since it's not problematic to install munge.pc even when
pkgconfig is not installed, this commit removes the check for the
"$(libdir)/pkgconfig" directory and always installs the munge.pc file.

Fixes issue 25.