Convert RPM specfile from sysvinit to systemd
This commit coverts the RPM specfile from sysvinit to systemd.
The requisite Requires & BuildRequires tags for systemd are added.
References to chkconfig and the sysvinit script are replaced with
corresponding helper macros to handle systemd scriptlet operations.
The sysvinit script and sysconfig file are replaced with the systemd
unit file and tmpfiles.d configuration file.

Furthermore, the %ghost tag is removed from the /var/run/munge
directory.  This tag was added back in commit 0f89429, at which
point the sysvinit script created this directory when the service
was started.  By removing the %ghost tag, this directory will be
created when the rpm is installed.  Since /var/run is (likely)
mounted as tmpfs, this directory will disappear on the next reboot,
at which point it will be re-created by tmpfiles.d.

The RuntimeDirectory and RuntimeDirectoryMode options in the service
file were added in systemd 211.  This is a clean and straightforward
manner in which to create private runtime directories below /run.
However, tmpfiles.d is being used instead since CentOS 7 ships with
systemd 208.

References:
- https://fedoraproject.org/wiki/Packaging:Systemd
- https://fedoraproject.org/wiki/Packaging:ScriptletSnippets
- https://fedoraproject.org/wiki/Packaging:Tmpfiles.d

Tested on:
- CentOS 7
- Fedora 20-23

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>
Closes #33
Closes #43
Issue #27
Issue #36
Issue #39