Update RPM specfile
Change the Source0 tag to use the full URL to the distributed tarball.

Remove the BuildRoot tag since it is ignored as of Fedora 18.

Change the Requires tag in the devel subpackage to use the %{?isa}
macro in order to make the dependency arch-specific.

Remove the static library (and configure with "--disable-static").
However, the libtool archive is being retained since removing it
constitutes an API/ABI change.

Remove rpath.

Change the invocation of make to use %{?_smp_mflags} for parallel make.

Replace "$RPM_BUILD_ROOT" with "%{buildroot}".

Change the login shell for the munge account created via useradd.

Add "exit 0" to the end of the %pre scriptlet used for dynamic
allocation of the munge user/group.  This exit allows the scriptlet
to succeed even if the munge user/group creation fails.

Remove %post stanzas that dealt with upgrading to 0.5.11.

Change the "%post libs" and "%postun libs" scriptlets to use the "-p"
option in order to run the ldconfig command without invoking a shell.

Remove %defattr at the beginning of each %files section since these
permissions are the default as of rpm 4.4.

Change the tag for COPYING and COPYING.LESSER from %doc to %license.
Also add conditional test for _licensedir to define %license if needed.

References:
- https://fedoraproject.org/wiki/How_to_create_an_RPM_package
- https://fedoraproject.org/wiki/Packaging:Guidelines
- https://fedoraproject.org/wiki/Packaging:LicensingGuidelines
- https://fedoraproject.org/wiki/Packaging:UsersAndGroups
- https://fedoraproject.org/wiki/EPEL:Packaging#The_.25license_tag

Tested on:
- CentOS 7
- Fedora 20-23

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>