Ignore group-writable permissions on LogDir
Starting in Ubuntu 14.04, the permissions and ownership on the
/var/log directory changed from 0755 root:root to 0775 root:syslog.
This was done to allow rsyslogd to dynamically create files in that
directory after having dropped privileges via "$PrivDropToUser syslog"
and "$PrivDropToGroup syslog".

The munged daemon enforces secure permissions and ownership on
its files and directories.  These directories must be owned by the
user that the munged daemon will run as.  They cannot allow write
permissions for group or other (unless the sticky-bit is set).
In addition, all of their parent directories in the path on up to
the root directory must be owned by either root or the user that the
munged daemon will run as.  None of these directories can allow write
permissions for group or other (unless the sticky-bit is set).

With the group-writable permission set on the /var/log directory,
munged fails to start with the following error:

  munged: Error: Logfile is insecure: group-writable permissions set on "/var/log"

The directory checks are performed by path_is_secure().  This routine
ensures the base directory cannot be modified by anyone other than
the current user or root.  It is called from multiple places within
munged to check various paths.

This commit adds a "flags" arg to path_is_secure(), and declares
the path_security_flag_t enum for setting flags to control which
security checks are performed.  The PATH_SECURITY_IGNORE_GROUP_WRITE
flag disables the group-writable permission check on the directory.
This allows the group-writable permission check to be disabled on
the LogDir while still enabling it for all other directories used
by munged.

References:
- https://bugs.launchpad.net/ubuntu/+source/munge/+bug/1287624
- https://bugs.launchpad.net/ubuntu/+source/rsyslog/+bug/484336
- https://bugs.launchpad.net/ubuntu/+source/rsyslog/+bug/1256695
- http://www.rsyslog.com/doc/droppriv.html

Tested on:
- CentOS 7
- Fedora 23
- Ubuntu 15.10

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>
Closes #31