Add munged "--max-ttl" cmdline opt
When munged decodes a credential, it must maintain this state until the
credential has expired in order to prevent it from being replayed.
Decoded credentials are stored in the replay cache.  This cache
is implemented as a chained hash.  Expired credentials are purged
every 60 seconds (MUNGE_REPLAY_PURGE_SECS).  If munged is decoding
credentials at a sustained high rate, the replay cache will consume
an increasing amount of memory with decreasing performance.

This commit adds the "--max-ttl" command-line option to munged.
This setting preserves the upper-bound imposed by MUNGE_MAXIMUM_TTL
(3600).  Reducing it will limit the maximum growth of the credential
replay cache.  This is viable if clocks within the MUNGE realm can
be kept in sync with minimal skew.

Tested on:
- CentOS 7
- Debian 8
- Fedora 23
- FreeBSD 10.2
- OpenBSD 5.8
- openSUSE Leap 42.1
- Ubuntu 15.10

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>
Closes #28