Add support for "make dist" and "make distcheck"
This adds support for the makefile "dist" and "distcheck" targets.

Automake "*_SOURCES" and "EXTRA_DIST" variables are updated to capture
all sources needed for "dist".

The "uninstall-local" target is added to src/etc/Makefile.am for
"distcheck" to succeed.

The X_AC_META macro is updated to set "PACKAGE" and "VERSION" from
the META file since these variables are needed by the "dist" targets.

AUTOMAKE_OPTIONS are updated to include "dist-xz" and "no-dist-gzip".
This causes "make dist" to generate a single tar.xz archive.
The tar.xz archive is necessary because commit 8cd3e16 changed the
munge.spec Source0 tag to use a tarball name with the ".xz" suffix.
Note that a "dist-bzip2" or "dist-gzip" generated tarball cannot be
directly passed to rpmbuild since the munge.spec Source0 tag will
fail to match the tarball name.

On CentOS 7 and Fedora 23, an rpm package can be built as follows:
  $ ./configure && make distcheck && rpmbuild -ta --clean *.tar.xz

Tested on:
- CentOS 7
- Fedora 23

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>
Closes #45