Fix enum path_security_flags -Wpedantic warning
C99 allows a comma to be placed at the end of an enumeration definition
list.  But "--enable-debug" adds "-Wall -pedantic", and the trailing
comma in enum path_security_flags triggers the following warning:

  path.h:47:45: warning: comma at end of enumerator list [-Wpedantic]
       PATH_SECURITY_IGNORE_GROUP_WRITE =  0x01,
                                               ^

This commit removes the trailing comma to appease -Wpedantic and be
consistent with the rest of the codebase.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>