Ensure daemonize_fini() leaves fd=0,1,2 open
In daemonize_fini(), /dev/null is open()d onto a temporary file
descriptor (stored in "dev_null") which is then dup()d onto stdin,
stdout, and stderr.  But if file descriptors 0, 1, or 2 are not
already open at the start of this function, then the temporary file
descriptor will fall within this range and be inadvertently closed
when the function cleans up.

This commit adds a check to prevent inadvertently closing this
(presumably temporary) file descriptor when it shouldn't.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>