Remove libmissing dependency from libmunge
libmunge has dependencies on two libtool convenience libraries,
libcommon and libmissing.  The shared library libmunge.so exports
only those symbols in its public interface, but the static archive
libmunge.a cannot restrict access to its objects.  This has recently
caused a problem with an application statically linking with both
liblustreapi.a and libmunge.a; both of these static archives contain
the strlcpy symbol.

In libmunge, the only routine used from libmissing is strlcpy(),
and it is used in only one place.

This commit replaces that single instance of strlcpy() with strncpy(),
and removes libmissing.la from libmunge.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>
Closes #49
Closes #50