Setup Travis CI builds
This commit enables Travis CI continuous integration builds at
<https://travis-ci.org/>.

A build matrix is defined to create debug and production builds
with libgcrypt and openssl cryptographic libraries using gcc and
clang compilers.

The "-Werror" CFLAGS option is specified for all builds in order
to change warnings into errors.  Note that neither "-pedantic" nor
"-pedantic-errors" is specified in CFLAGS.  libtool will fail if either
is used in the CFLAGS passed to the configure script.  It should be
possible to override libtool's CFLAGS via LTCFLAGS, but this does
not work; LTCFLAGS always takes on the value of CFLAGS.  Grrr.

DISTCHECK_CONFIGURE_FLAGS is needed to pass CONFIGURE_OPTS
to the configure script in "make distcheck".  Note that the
DISTCHECK_CONFIGURE_FLAGS value of CONFIGURE_OPTS is single-quoted
in the global environment definition.  This is necessary since
CONFIGURE_OPTS varies throughout the build matrix, and single-quoting
delays variable expansion until the "script" step of the build.

colormake is used to colorize the output of the Travis CI job log.

Reference:
- https://docs.travis-ci.com/user/getting-started/
- https://www.gnu.org/software/automake/manual/html_node/Checking-the-Distribution.html

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>