Fix Coverity Scan with Travis CI build matrix
The Travis CI config specifies a build matrix that expands to 8 jobs.
When running a Coverity Scan (by pushing to the coverity_scan branch),
each job attempts to run a scan on the same git tree.  This not only
wastes resources, but quickly overruns the scan quota.

This commit adds a "before_install" step that immediately exits if
the branch is the "coverity_scan" branch and the job number is not 1.
This also modifies the "script" to omit the normal build step on the
"coverity_scan" branch.

Reference:
- https://github.com/travis-ci/travis-ci/issues/1975#issuecomment-216604414

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>