Revisit read_data_from_file & read_data_from_string
Coverity Scan identified an overflowed/truncated value in
read_data_from_file():

  n = bufptr - bufmem;

Here, 'n' is a signed integer, and 'bufptr' and 'bufmem' are both
unsigned char pointers.  When this code was written (back on ilp32),
these were all 32-bit values.  But on i32lp64, the integer is 32-bits
while the pointers are 64-bits; also, since the integer is signed,
its maximum value is (2^31 - 1).  This results in a 64-bit value
being assigned to a 32-bit variable.  Pragmatically, this isn't much
of a concern since (bufptr - bufmem) will be a relatively small value
that can fit within a signed integer.  But while I'm poking at it...

 1. This commit changes the local signed int variables to type size_t.

 2. Since the function prototype defines 'len' as a signed integer
    pointer, it checks to ensure the size_t buffer length value can
    be stored in a signed int variable.

 3. It sets a MAXIMUM_BUFFER_SIZE limit on the buffer size since
    messages exceeding the MUNGE_MAXIMUM_REQ_LEN limit over the unix
    domain socket will be rejected.

 4. It removes the final realloc() which shrinks the memory allocation
    of the buffer to (buflen + 1).  This seems unnecessary considering
    the buffer is short-lived and will likely use the same amount of
    physical memory.

 5. It changes the return type from int to void.  Before this change,
    the function would return -1 on error, after which the caller would
    log a fatal error message.  Now, it can log a fatal error message
    having a more nuanced description of the error.  All callers
    (i.e., munge.c:main and unmunge.c:main) are updated accordingly.

 6. It similarly changes the return type and error handling of
    read_data_from_string().

 7. It restructures and comments the code for clarity.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>
Coverity ID 1363054