Adjust socket timeouts
For the client's delay between attempts at connecting to the Unix
domain socket, this commit reduces the starting point of the linear
backoff (MUNGE_SOCKET_CONNECT_RETRY_MSECS) to 50ms and increases the
number of attempts (MUNGE_SOCKET_CONNECT_ATTEMPTS) to 10.

It also decreases the timeout for reads/writes on the socket
(MUNGE_SOCKET_TIMEOUT_MSECS) from 3000ms to 2000ms.

Finally, this commit removes the comment and assertion that
(MUNGE_SOCKET_RETRY_ATTEMPTS * MUNGE_SOCKET_RETRY_MSECS) < 1000.
The rationale for this stipulation no longer seems to apply.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>