Add munged "--log-file" & "--seed-file" cmdline opts
This commit adds two command-line options to munged: "--log-file"
and "--seed-file".  The former specifies an alternate pathname to
the daemon's log file, and the latter specifies an alternate pathname
to the PRNG seed file.  With the addition of these two options, all
MUNGE-specific compile-time paths can be overridden on the command-line
at runtime.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>
Closes #57