Rename functions in xgetgr.c & xgetpw.c
A portability abstraction for getgrnam_r() would need routines to
perform a deep copy of the data from getgrnam()'s static buffer in
the same manner as xgetgrent() & xgetpwnam().  Instead of duplicating
the private routines in xgetgr.c, it makes sense to generalize and
restructure the routines in xgetgr.c (formerly xgetgrent.c) and
xgetpw.c (formerly xgetpwnam.c).

The following functions in xgetgr.c are renamed:

- xgetgrent_buf_create() -> xgetgrbuf_create()
- xgetgrent_buf_destroy() -> xgetgrbuf_destroy()
- xgetgrent_buf_get_len() -> xgetgrbuf_get_len()

- _xgetgrent_buf_get_sys_size() -> _xgetgrbuf_get_sys_size()
- _xgetgrent_buf_grow() -> _xgetgrbuf_grow()
- _xgetgrent_copy() -> _xgetgrbuf_copy_struct()
- _xgetgrent_copy_str() -> _xgetgrbuf_copy_string()

The following functions in xgetpw.c are renamed:

- xgetpwnam_buf_create() -> xgetpwbuf_create()
- xgetpwnam_buf_destroy() -> xgetpwbuf_destroy()
- xgetpwnam_buf_get_len() -> xgetpwbuf_get_len()

- _xgetpwnam_buf_get_sys_size() -> _xgetpwbuf_get_sys_size()
- _xgetpwnam_buf_grow() -> _xgetpwbuf_grow()
- _xgetpwnam_copy() -> _xgetpwbuf_copy_struct()
- _xgetpwnam_copy_str() -> _xgetpwbuf_copy_string()

All callers are updated.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>