Fix xgetgrent/xgetgrnam/xgetpwnam uninit'd rv_copy
If a re-entrant version of getgrent_r()/getgrnam_r()/getpwnam_r()
does not exist on the system, and the non-reentrant version of
getgrent()/getgrnam()/getpwnam() fails, then the variable rv_copy
would be uninitialized when it is accessed.  This could result in the
routine returning an error before it would otherwise ... return an
error (although in the case of ERANGE, this could prevent the call
from being retried with a larger buffer).

This commit fixes the 3 identical instances of the uninitialized
variable rv_copy in xgetgrent(), xgetgrnam(), and xgetpwnam().

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>