Revisit crypto_init for Libgcrypt
This revisits the crypto_init() implementation used with Libgcrypt
(i.e., surrounded by the HAVE_LIBGCRYPT preprocessor macro).

It adds a comment noting that the GCRYCTL_SET_THREAD_CBS command for
gcry_control() was marked obsolete starting in Libgcrypt 1.6.

It also changes the call to gcry_check_version() to perform the
version check as recommended in the Libgcrypt documentation.
This guards against problems with the dynamic linker using an old
version of the library with a newer version of the include header.

Reference:
- https://www.gnupg.org/documentation/manuals/gcrypt/Initializing-the-library.html

Tested on:
- Debian 3.1 (Libgcrypt 1.2.0)
- Debian 4.0 (Libgcrypt 1.2.3)
- Debian 5.0.10 (Libgcrypt 1.4.1)
- Debian 6.0.10 (Libgcrypt 1.4.5)
- Debian 7.11 (Libgcrypt 1.5.0)
- Debian 8.8 (Libgcrypt 1.6.3)
- Debian 9.0 (Libgcrypt 1.7.6-beta)

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>