Track hash_node memory allocations on hash_mem_list
Memory for the hash_node struct is allocated in blocks of 1024 objects
which are broken up and added to the hash_free_list.  This memory
is not explicitly deallocated which results in Valgrind's leak-check
reporting these allocations as being "possibly lost".

Replacing these block allocations with individual calls to malloc()
and free() for each object used noticeably more memory and slightly
more CPU time (resulting in decreased throughput according to the
remunge benchmark).

Instead, this commit begins each block allocation with a pointer for
chaining these allocations together on the hash_mem_list.  It also
adds the hash_drop_memory() routine to explicitly free this memory.
This routine is called from main() after the replay and gids hashes
have been destroyed.

In addition, this commit renames hash_free_lock to hash_free_list_lock,
and HASH_ALLOC to HASH_NODE_ALLOC_NUM.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>