Add support for getentropy and getrandom
Add support for getentropy() and getrandom() to
_random_read_entropy_from_kernel() for seeding the PRNG.

The rationale for these calls is to provide resilience against file
descriptor exhaustion attacks where the attacker consumes all available
file descriptors, thereby preventing access to entropy sources such
as /dev/urandom or a seed file.  These new calls can work in a chroot
or when file descriptors are exhausted.  Another benefit is that
these calls will block until the urandom pool has been initialized,
after which reads of up to 256 bytes will always return as many bytes
as requested.

getentropy() fills a buffer with high-quality entropy suitable for
seeding a user-space PRNG.  getrandom() also fills a buffer with
high-quality entropy; this can be used to seed a user-space PRNG or
for cryptographic purposes.

Darren Moffat writes that on Solaris, data returned by getentropy()
has not had the required FIPS 140-2 processing for the DRBG applied to
it; consequently, it should not be used directly where an IV or nonce
is needed when calling a cryptographic operation -- it is intended
only for seeding a user-space random bit generator.  In contrast,
the output of getrandom() on Solaris has been through a FIPS 140-2
approved DRBG function as defined in NIST SP-900-90A.

getentropy() first appeared in OpenBSD 5.6.  getrandom() was introduced
in version 3.17 of the Linux kernel.  Solaris 11.3 added support
for both getentropy() and getrandom().  Recently, glibc 2.25 added
support for both as well; Fedora 26 includes glibc 2.25.

Reference:
- https://lwn.net/Articles/711013/
- https://lwn.net/Articles/606141/
- https://man.openbsd.org/getentropy.2
- https://blogs.oracle.com/darren/solaris-new-system-calls:-getentropy2-and-getrandom2
- https://fedoraproject.org/wiki/Changes/GLIBC225

Tested on:
- CentOS 7.3.1611 (/dev/urandom)
- Fedora 26 (getrandom)
- OpenBSD 6.1 (getentropy)

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>