Remove RAND_load_file and RAND_write_file
During testing on OpenBSD 6.1 (with LibreSSL 2.5.2), it was discovered
that the PRNG was supposedly being seeded with 1024 bytes from
"munge.seed" despite that file not yet existing.  This bug was
traced to a change in the behavior of RAND_load_file() in LibreSSL.
This routine is used to seed the PRNG using the contents of the
specified file; it returns the number of bytes read.

In LibreSSL, RAND_load_file() is kept for ABI compatibility but is no
longer functional.  According to the man page, it always returns 0.
However, it seems to return the value of the "max_bytes" argument
regardless of whether "filename" even exists.  This behavior was
observed all the way back to OpenBSD 5.6 which introduced LibreSSL 2.0.

On a related note, the corresponding OpenSSL function RAND_write_file()
has a limitation in that it always writes 1024 bytes to the specified
filename; it does not support an argument to specify the number of
bytes to write.

Compatible routines were previously developed when adding support for
Libgcrypt.  Those routines do not exhibit these bugs or limitations.

This commit removes the OpenSSL routines for _random_read_seed() and
_random_write_seed(); the former was a wrapper for RAND_load_file(),
while the latter called RAND_write_file().  Instead, the Libgcrypt
implementation of these routines is moved out of that #ifdef-block
and altered to work for both Libgcrypt and OpenSSL.

Tested on:
- CentOS 7.3.1611 (OpenSSL 1.0.1e-fips 11 Feb 2013)
- Debian 9.0 (Libgcrypt 1.7.6-beta)
- Fedora 26 (OpenSSL 1.1.0f-fips 25 May 2017)
- OpenBSD 5.6 (LibreSSL 2.0)
- OpenBSD 5.7 (LibreSSL 2.1)
- OpenBSD 5.8 (LibreSSL 2.2.2)
- OpenBSD 5.9 (LibreSSL 2.3.2)
- OpenBSD 6.0 (LibreSSL 2.4.2)
- OpenBSD 6.1 (LibreSSL 2.5.2)

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>